refactor: move to fs/utils.ts
