+++ test.ts
@@ -20,6 +20,7 @@ import "./fs/ensure_file_test.ts";
 import "./fs/move_test.ts";
 import "./fs/read_json_test.ts";
 import "./fs/write_json_test.ts";
+import "./fs/utils_test.ts";
 import "./io/test.ts";
 import "./http/server_test.ts";
 import "./http/file_server_test.ts";
